refactored code and switched to huggingface model google/flan-t5-large. Not the best one, but it works
