+++ utils/document_handler.py
@@ -1,19 +1,8 @@
 import os
 import pathlib
-import shutil
 
-from langchain.chains import RetrievalQA, ConversationalRetrievalChain
-from langchain.chains.query_constructor.base import AttributeInfo
 from langchain.document_loaders import DirectoryLoader
-from langchain.embeddings.openai import OpenAIEmbeddings
-from langchain.llms import OpenAI
-from langchain.memory import ConversationBufferMemory
-from langchain.prompts import PromptTemplate
-from langchain.retrievers import ContextualCompressionRetriever
-from langchain.retrievers.document_compressors import LLMChainExtractor
-from langchain.retrievers.self_query.base import SelfQueryRetriever
 from langchain.text_splitter import RecursiveCharacterTextSplitter
-from langchain.vectorstores import Chroma
 
 
 class TextLoader:
@@ -65,154 +54,3 @@ class TextLoader:
     def get_splits(self):
         self.splits = self.splitter.split_documents(self.corpus)
         return self.splits
-
-
-class VectorDatabase:
-    def __init__(self,
-                 vector_db_path: pathlib.Path,
-                 embeddings=OpenAIEmbeddings()):
-        """
-        Parameters
-        ----------
-        vector_db_path:
-            path to the vector database, should be the same as the book path
-        """
-        self.vector_store_dir = str(vector_db_path.joinpath('database'))
-        self.embeddings = embeddings
-
-        self.vector_db = self.set_vector_store()
-
-        self.set_llm('open-ai')
-        self.set_meta_data_fields()
-        self.set_retriever()
-
-    def create_vector_store(self, splits, delete_existing_db=True):
-        if delete_existing_db and os.path.isdir(self.vector_store_dir):
-            shutil.rmtree(self.vector_store_dir)
-
-        self.vector_db = Chroma.from_documents(
-                documents=splits,
-                embedding=self.embeddings,
-                persist_directory=self.vector_store_dir
-        )
-        self.vector_db.persist()
-
-    def set_vector_store(self):
-        vector_db = Chroma(
-                persist_directory=self.vector_store_dir,
-                embedding_function=self.embeddings
-        )
-        # ToDo: add exception, if DB does not exist
-        return vector_db
-
-    def set_llm(self, llm_type='open-ai'):
-        if llm_type == 'open-ai':
-            self.llm = OpenAI(temperature=0.1)
-        elif llm_type == 'open-source':
-            raise NotImplementedError('Open source alternative is not '
-                                      'implemented yet.')
-        else:
-            raise ValueError(f'llm_type {llm_type} is unknown.\n'
-                             f'Valid arguments are: "open_ai" or '
-                             f'"open-source"')
-
-    def set_meta_data_fields(self):
-        query_metadata = [
-            AttributeInfo(
-                    name="source",
-                    description="The raw data, which contains the book.",
-                    type="string",
-            ),
-            AttributeInfo(
-                    name="chapter",
-                    description="The chapter from the book",
-                    type="integer",
-            ),
-            AttributeInfo(
-                    name="subchapter",
-                    description="The subchapter is part of the chapter and "
-                                "therefore hierarchically lower.",
-                    type="integer",
-            ),
-        ]
-        self.query_metadata = query_metadata
-
-    def set_retriever(self, retriever_type='self'):
-        if retriever_type == 'self':
-            document_content_description = "Book"
-            retriever = SelfQueryRetriever.from_llm(
-                    self.llm,
-                    self.vector_db,
-                    document_content_description,
-                    self.query_metadata,
-                    verbose=True
-            )
-        elif retriever_type == 'compress':
-            compressor = LLMChainExtractor.from_llm(self.llm)
-            retriever = ContextualCompressionRetriever(
-                    base_compressor=compressor,
-                    base_retriever=self.vector_db.as_retriever()
-            )
-        self.retriever = retriever
-
-
-class AnswerMe:
-    def __init__(self, llm, vector_db, retriever):
-        self.llm = llm
-        self.vector_db = vector_db
-        self.retriever = retriever
-
-        self.set_template()
-        self.set_memory()
-        self.set_conversation_answerer()
-        
-        # self.set_answerer()
-
-    def get_relevant_documents(self, question):
-        return self.retriever.get_relevant_documents(query=question)
-
-    def set_answerer(self):
-        """
-        Does not have context capacity, therefore will not be used now
-        Returns
-        -------
-
-        """
-        prompt_template = PromptTemplate.from_template(self.template)
-        self.answerer = RetrievalQA.from_chain_type(
-                self.llm,
-                retriever=self.vector_db.as_retriever(search_type="mmr"),
-                # return_source_documents=True,
-                chain_type_kwargs={'prompt': prompt_template},
-                chain_type='stuff'
-        )
-
-    def answer(self, question):
-        answer = self.answerer({'question': question})
-        return answer['answer']
-
-    def set_template(self):
-        self.template = \
-            """
-            Answer like a literature expert. Only use the context provided to 
-            answer the given question, don't try to make up an answer. 
-            Mention in which chapter or 
-            subchapter the answer can be looked up.
-            {context}
-            Question: {question}
-            Answer:"""
-
-    def set_memory(self):
-        self.memory = ConversationBufferMemory(
-                memory_key="chat_history",
-                return_messages=True
-        )
-
-    def set_conversation_answerer(self):
-        prompt_template = PromptTemplate.from_template(self.template)
-        self.answerer = ConversationalRetrievalChain.from_llm(
-                self.llm,
-                retriever=self.retriever,
-                memory=self.memory,
-                combine_docs_chain_kwargs={'prompt': prompt_template}
-        )
